// types/SystemData.t.ds.ts
export interface SystemData {


    name: string;
    description: string;
    code: string;
    affiliation: string;
    type: string;
    subtype: string;
    status: string;
    frost_line: string;
    product_thumb_large: string;
    position_x: number;
    position_y: number;
    position_z: number;
    data: any;
}